# docs/devicemanager_swagger.yaml
basePath: /
definitions:
...
  models.Attribute:
    properties:
      key:
        type: string
      origin:
        type: string
      value:
        type: string
    type: object
...
  models.Content:
    properties:
      content_variable:
        $ref: '#/definitions/models.ContentVariable'
      id:
        type: string
      protocol_segment_id:
        type: string
      serialization:
        $ref: '#/definitions/models.Serialization'
    type: object
  models.ContentVariable:
    properties:
      aspect_id:
        type: string
      characteristic_id:
        type: string
      function_id:
        type: string
      id:
        type: string
      is_void:
        type: boolean
      name:
        type: string
      omit_empty:
        type: boolean
      serialization_options:
        items:
          type: string
        type: array
      sub_content_variables:
        items:
          $ref: '#/definitions/models.ContentVariable'
        type: array
      type:
        $ref: '#/definitions/models.Type'
      unit_reference:
        type: string
      value: {}
    type: object
...
  models.DeviceType:
    properties:
      attributes:
        items:
          $ref: '#/definitions/models.Attribute'
        type: array
      description:
        type: string
      device_class_id:
        type: string
      id:
        type: string
      name:
        type: string
      service_groups:
        items:
          $ref: '#/definitions/models.ServiceGroup'
        type: array
      services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
    type: object
  models.Function:
    properties:
      concept_id:
        type: string
      description:
        type: string
      display_name:
        type: string
      id:
        type: string
      name:
        type: string
      rdf_type:
        type: string
    type: object
...
  models.Interaction:
    enum:
    - event
    - request
    - event+request
    type: string
    x-enum-varnames:
    - EVENT
    - REQUEST
    - EVENT_AND_REQUEST
...
  models.Serialization:
    enum:
    - xml
    - json
    - plain-text
    type: string
    x-enum-varnames:
    - XML
    - JSON
    - PlainText
  models.Service:
    properties:
      attributes:
        items:
          $ref: '#/definitions/models.Attribute'
        type: array
      description:
        type: string
      id:
        type: string
      inputs:
        items:
          $ref: '#/definitions/models.Content'
        type: array
      interaction:
        $ref: '#/definitions/models.Interaction'
      local_id:
        type: string
      name:
        type: string
      outputs:
        items:
          $ref: '#/definitions/models.Content'
        type: array
      protocol_id:
        type: string
      service_group_key:
        type: string
    type: object
  models.ServiceGroup:
    properties:
      description:
        type: string
      key:
        type: string
      name:
        type: string
    type: object
  models.Type:
    enum:
    - https://schema.org/Text
    - https://schema.org/Integer
    - https://schema.org/Float
    - https://schema.org/Boolean
    - https://schema.org/ItemList
    - https://schema.org/StructuredValue
    type: string
    x-enum-varnames:
    - String
    - Integer
    - Float
    - Boolean
    - List
    - Structure
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Device-Manager API
  version: "0.1"
paths:
...
  /device-types:
    post:
      description: create device-type
      parameters:
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      - description: comma separated list of attribute keys; no other device-type
          with the same attribute key/value may exist
        in: query
        name: distinct_attributes
        type: string
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.DeviceType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeviceType'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: create device-type
      tags:
      - create
      - device-types
  /device-types/{id}:
    delete:
      description: get device-type
      parameters:
      - description: DeviceType Id
        in: path
        name: id
        required: true
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: get device-type
      tags:
      - get
      - device-types
    get:
      description: get device-type
      parameters:
      - description: DeviceType Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeviceType'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: get device-type
      tags:
      - get
      - device-types
    put:
      description: set device-type
      parameters:
      - description: DeviceType Id
        in: path
        name: id
        required: true
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      - description: comma separated list of attribute keys; no other device-type
          with the same attribute key/value may exist
        in: query
        name: distinct_attributes
        type: string
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.DeviceType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeviceType'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: set device-type
      tags:
      - set
      - device-types
...
  /functions:
    post:
      description: create function
      parameters:
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Function'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Function'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: create function
      tags:
      - create
      - functions
  /functions/{id}:
    delete:
      description: delete function
      parameters:
      - description: Function Id
        in: path
        name: id
        required: true
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: delete function
      tags:
      - delete
      - functions
    get:
      description: get function
      parameters:
      - description: Function Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Function'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: get function
      tags:
      - get
      - functions
    put:
      description: set function
      parameters:
      - description: Function Id
        in: path
        name: id
        required: true
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Function'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Function'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: set function
      tags:
      - set
      - functions
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"